% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTLo.R
\name{dLTLo}
\alias{dLTLo}
\alias{sLTLo}
\alias{hLTLo}
\alias{pLTLo}
\alias{qLTLo}
\alias{rLTLo}
\title{The Logarithmic Transformed Lomax}
\usage{
dLTLo(x, phi, sigma, tau = 1, log = FALSE)

sLTLo(x, phi, sigma, tau = 1, log = FALSE)

hLTLo(x, phi, sigma, tau = 1, log = FALSE)

pLTLo(q, phi, sigma, tau, lower.tail = TRUE, log.p = FALSE)

qLTLo(q, phi, sigma, tau, lower.tail = TRUE, log.p = FALSE)

rLTLo(n, phi, sigma, tau)
}
\arguments{
\item{\code{x, q}}{vector of quantiles.}

\item{\code{p}}{vector of probabilities.}

\item{\code{n}}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{\code{phi, sigma}}{vectors of shape parameters.}

\item{\code{tau}}{vector of scale parameters.}

\item{\code{log, log.p}}{logical; if TRUE, probabilities p are given as log(p).}

\item{\code{lower.tail}}{logical; if TRUE (default), probabilities are P\link{X≤x}, otherwise, P\link{X>x}.}
}
\value{
\code{dLTLo} gives the density, \code{sLTLo} gives the survival function, \code{hLTLo} gives the hazard rates function, \code{pLTLo} gives the distribution function, \code{qLTLo} gives the quantile function, and \code{rLTLo} generates random deviates.

The length of the result is determined by \code{n} for \code{rLTo}, and is the maximum of the lengths of the numerical arguments for the other functions.

The numerical arguments other than \code{n} are recycled to the length of the result. Only the first elements of the logical arguments are used.
}
\description{
Density, survival function, hazard rate function, distribution function, quantile function and random generation for
the Logarithmic Transformed Lomax distribution. This distribution is an extension of the traditional Lomax distribution
with one scale parameter and two shape parameters.
}
\details{
If \eqn{\tau} is not specified, it assumes the default value of 1.

The Logarithmic Transformed Lomax distribution has density

\deqn{f(x;\phi,\sigma,\tau) = \frac{(\phi-1)\sigma\tau^\sigma(\tau+x)^{-(\sigma+1)}}{\log(\phi)[1-(1-\phi)\tau^\sigma(\tau^\sigma(\tau+x)^{-\sigma})]}}

for \eqn{x,\phi,\sigma,\tau \geq 0}, \eqn{\phi \neq 1.}
}
\references{
Alotaibi, R., Okasha, H., Rezk, H., Almarashi, A. M., & Nassar, M. (2021).
On a new flexible Lomax distribution: statistical properties and estimation procedures
with applications to engineering and medical data. AIMS Mathematics, 6(12), 13976-13999.

Lomax, K. S. (1954). Business failures: Another example of the analysis of failure data. Journal of the American Statistical Association, 49(268), 847-852.
}
